+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/adapter/HandlerExecutionAdapter.java
@@ -1,6 +1,7 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.adapter;
 
-import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.mvc.handler.HandlerExecution;
+import com.interface21.webmvc.servlet.view.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
