+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/mapping/HandlerMappingRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.mapping;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
